Fix for `_schema_version` overriding non-object entity data (#137)
We recommend that Snowplow entities should be JSON *objects* only. But
technically it is possible for them to be something else, e.g. a JSON
array.

Since version 3.2.0 we started adding the `_schema_version` field to
context entities, see #132. But this accidentally clobbered the original
data if the entity was not an object. This commit fixes the problem by
only adding `_schema_version` if the entity is an object.